+++ print/src/main.rs
@@ -20,6 +20,12 @@ struct Opts {
     json: bool,
 
     #[arg(long)]
+    log_format: Option<String>,
+
+    #[arg(long)]
+    host_log_format: Option<String>,
+
+    #[arg(long)]
     show_skipped_frames: bool,
 
     #[arg(short, long)]
@@ -80,6 +86,8 @@ fn main() -> anyhow::Result<()> {
     let Opts {
         elf,
         json,
+        log_format,
+        host_log_format,
         show_skipped_frames,
         verbose,
         version,
@@ -90,10 +98,15 @@ fn main() -> anyhow::Result<()> {
         return print_version();
     }
 
-    defmt_decoder::log::init_logger(verbose, json, move |metadata| match verbose {
-        false => defmt_decoder::log::is_defmt_frame(metadata), // We display *all* defmt frames, but nothing else.
-        true => true,                                          // We display *all* frames.
-    });
+    defmt_decoder::log::init_logger(
+        log_format.as_deref(),
+        host_log_format.as_deref(),
+        json,
+        move |metadata| match verbose {
+            false => defmt_decoder::log::is_defmt_frame(metadata), // We display *all* defmt frames, but nothing else.
+            true => true,                                          // We display *all* frames.
+        },
+    );
 
     // read and parse elf file
     let bytes = fs::read(elf.unwrap())?;